requirements.txt
python-telegram-bot==13.15
pandas
numpy
matplotlib
mplfinance
ta
python-dotenv
requests